requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
chardet==3.0.4
dj-database-url==0.5.0
Django==3.0.6
django-cors-headers==3.2.1
django-heroku==0.3.1
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
gunicorn==20.0.4
idna==2.9
Pillow==7.1.2
psycopg2==2.8.5
PyJWT==1.7.1
python-decouple==3.0
pytz==2020.1
requests==2.23.0
sqlparse==0.3.1
urllib3==1.25.9
whitenoise==5.0.1
